% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{classInd}
\alias{classInd}
\title{Function for k-means}
\usage{
classInd(cl)
}
\arguments{
\item{cl}{Label list of nodes}
}
\value{
x    : class indicator matrix
}
\description{
Function for k-means
}
