% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{confidenceInterval}
\alias{confidenceInterval}
\title{Confidence Interval}
\usage{
confidenceInterval(boot.sol, alpha = 0.05)
}
\arguments{
\item{boot.sol}{Bootstrap list of estimators}

\item{alpha}{Level of confidence : 1 - \eqn{\alpha}}
}
\value{
Array with 3 dimensions and size \eqn{Q(Q+1)/2\times 3\times K} (if undirected) or \eqn{Q^2\times 3\times K} (when undirected)
containing for each pair of groups \eqn{(q,l)} (first dimension) and each \eqn{k}-th subinterval (third dimension) the 3 quantiles at levels \eqn{(\alpha/2,1-\alpha/2,.5)} (second dimension).
}
\description{
Compute confidence bands for all pair of groups \eqn{(q,l)}
}
\keyword{internal}
