% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{correctTau}
\alias{correctTau}
\title{Handling of values of \eqn{\tau}}
\usage{
correctTau(tau)
}
\arguments{
\item{tau}{\eqn{\tau}}
}
\description{
Avoid values of \eqn{\tau} to be exactly 0 and exactly 1.
}
