% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{find_ql_diff}
\alias{find_ql_diff}
\title{Convert index into group pair in tauDown_Q}
\usage{
find_ql_diff(ind_ql, Q)
}
\arguments{
\item{ind_ql}{Converted \eqn{(q,l)} index}

\item{Q}{Total number of groups \eqn{Q}}
}
\value{
Group pair \eqn{(q,l)} corresponding to the given index
}
\description{
This function is the inverse of the conversion \eqn{{(q,l), q<l}} into \eqn{{1,...,Q*(Q-1)/2}}. Used only in tauDown_Q.
}
