% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generatePPConst}
\alias{generatePPConst}
\title{Poisson process  with piecewise constant intensities}
\usage{
generatePPConst(intens, Time)
}
\arguments{
\item{intens}{Vector with the constants of the intensities (defined on a regular partition of interval [0,Time])}

\item{Time}{Time}
}
\description{
Generate realizations of a Poisson process with piecewise constant intensities
}
\examples{
intens <- c(1,3,8)
constpp <- generatePPConst(intens, 10)

}
