% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{JEvalMstep}
\alias{JEvalMstep}
\title{Evaluation of criterion J}
\usage{
JEvalMstep(VE, mstep, data, directed, sparse, method = "hist")
}
\arguments{
\item{VE}{Results of the previous VE for iterative computation}

\item{mstep}{Results of the previous mstep for iterative computation
\itemize{
  \item mstep$sum_rhotau : N_Q vector (not needed in the function)
  \item mstep$sum_rhotau_obs : N_Q vector
  \item mstep$logintensities.ql :  N_Q x Dmax matrix
  \item m.step$beta : N_Q vector
}}

\item{data}{Data same of \link[ppsbm]{mainVEM}}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}

\item{sparse}{Boolean for sparse (TRUE) or not sparse (FALSE) case}

\item{method}{List of string. Can be "hist" for histogram method or "kernel" for kernel method}
}
\description{
Evaluation of the criterion J to verify the convergence of the VEM algorithm
}
