% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{listNodePairs}
\alias{listNodePairs}
\title{List node pairs}
\usage{
listNodePairs(n, directed = TRUE)
}
\arguments{
\item{n}{Total number of nodes}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}
}
\value{
Matrix with two columns which lists all the possible node pairs. Each row is a node pair.
}
\description{
Create the list of all node pairs
}
\examples{
# List all the node pairs with 10 nodes, for directed and undirected graphs

n <- 10
listNodePairs(n, TRUE)
listNodePairs(n, FALSE)

}
