% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{mainVEMPar}
\alias{mainVEMPar}
\title{VEM step for parallel version}
\usage{
mainVEMPar(
  init.point,
  n,
  Q,
  data,
  directed,
  sparse,
  method,
  nb.iter,
  fix.iter,
  epsilon
)
}
\arguments{
\item{init.point}{Initial point}

\item{n}{Total number of nodes,  \eqn{1\le i \le n}}

\item{Q}{Total number of groups}

\item{data}{Data same of \link[ppsbm]{mainVEM}}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}

\item{sparse}{Boolean for sparse (TRUE) or not sparse (FALSE) case}

\item{method}{List of string. Can be "hist" for histogram method or "kernel" for kernel method}

\item{nb.iter}{Number of iterations}

\item{fix.iter}{Maximum number of iterations of the fixed point}

\item{epsilon}{Threshold for the stopping criterion of VEM and fixed point iterations}
}
\description{
VEM step for parallel version
}
\keyword{internal}
