% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{modelSelec_QPlot}
\alias{modelSelec_QPlot}
\title{Plots for model selection}
\usage{
modelSelec_QPlot(model.selec_Q)
}
\arguments{
\item{model.selec_Q}{Output from modelSelection_Q()}
}
\description{
Plots for model selection
}
\examples{
# load data of a synthetic graph with 50 individuals and 3 clusters
n <- 50

# compute data matrix with precision d_max=3
Dmax <- 2^3
data <- list(Nijk=statistics(generated_Q3$data,n,Dmax,directed=FALSE),
    Time=generated_Q3$data$Time)

# ICL-model selection
sol.selec_Q <- modelSelection_Q(data,n,Qmin=1,Qmax=4,directed=FALSE,
    sparse=FALSE,generated_sol_hist)

# plot ICL
modelSelec_QPlot(sol.selec_Q)

}
