% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{Mstep_hist}
\alias{Mstep_hist}
\title{M step for histograms}
\usage{
Mstep_hist(data, VE, directed, sparse)
}
\arguments{
\item{data}{Data same of \link[ppsbm]{mainVEM}}

\item{VE}{Results of the previous VE for iterative computation}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}

\item{sparse}{Boolean for sparse (TRUE) or not sparse (FALSE) case}
}
\description{
M step for histograms estimator
}
\references{
BARAUD, Y. & BIRGÉ, L. (2009). Estimating the intensity of a random measure by histogram type estimators. Probab. Theory Related Fields 143, 239–284.

MATIAS, C., REBAFKA, T. & VILLERS, F. (2018).  A semiparametric extension of the stochastic block model for longitudinal networks. Biometrika. 105(3): 665-680.

REYNAUD -BOURET, P. (2006). Penalized projection estimators of the Aalen multiplicative intensity. Bernoulli 12, 633–661.
}
\keyword{internal}
