% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{Mstep_kernel}
\alias{Mstep_kernel}
\title{M step for kernel}
\usage{
Mstep_kernel(data, VE, directed)
}
\arguments{
\item{data}{Data same of \link[ppsbm]{mainVEM}}

\item{VE}{Results of the previous VE for iterative computation}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}
}
\description{
M step for kernel estimator
}
\references{
GRÉGOIRE , G. (1993). Least squares cross-validation for counting process intensities. Scand. J. Statist. 20, pp. 343–360.

MATIAS, C., REBAFKA, T. & VILLERS, F. (2018).  A semiparametric extension of the stochastic block model for longitudinal networks. Biometrika. 105(3): 665-680.

RAMLAU-HANSEN, H. (1983). Smoothing counting process intensities by means of kernel functions. Ann. Statist. 11, pp. 453–466.
}
\keyword{internal}
