% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{taurhoInitial}
\alias{taurhoInitial}
\title{Sparse setup - \eqn{\rho} parameter}
\usage{
taurhoInitial(tau, data, n, Q, directed = TRUE)
}
\arguments{
\item{tau}{\eqn{\tau}}

\item{data}{Data : only needs the \eqn{N_{ijk}} field of data}

\item{n}{Total number of nodes}

\item{Q}{Total number of groups}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}
}
\value{
Both \eqn{\tau} and \eqn{\rho}.
}
\description{
Sparse setup - \eqn{\rho} parameter
}
\keyword{internal}
