% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{tauUp_Q}
\alias{tauUp_Q}
\title{Construct initial \eqn{\tau} from \eqn{Q-1}}
\usage{
tauUp_Q(tau, n_perturb = 1)
}
\arguments{
\item{tau}{\eqn{\tau}}

\item{n_perturb}{Number of different perturbations on k-means result}
}
\value{
List of matrixes of initial values for \eqn{\tau} for \eqn{Q} groups from value obtained at \eqn{Q-1}
}
\description{
Construct initial \eqn{\tau} with \eqn{Q} groups from value obtained at \eqn{Q-1} groups
}
\keyword{internal}
