% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{tauUpdate}
\alias{tauUpdate}
\title{Update \eqn{\tau}}
\usage{
tauUpdate(tau, pi, mstep, data, directed, sparse, method, rho)
}
\arguments{
\item{tau}{Old \eqn{\tau}}

\item{pi}{Estimator of group probabilities \eqn{\pi}}

\item{mstep}{Results of the previous mstep for iterative computation}

\item{data}{Data same of \link[ppsbm]{mainVEM}}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}

\item{sparse}{Boolean for sparse (TRUE) or not sparse (FALSE) case}

\item{method}{List of string. Can be \code{hist} for histogram method or \code{kernel} for kernel method}

\item{rho}{Old \eqn{\rho} (only for \code{sparse} model, set to 1 otherwise)}
}
\description{
One update of \eqn{\tau} by the fixed point equation
}
\keyword{internal}
