% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem_dynppsbm.R
\name{VEstep}
\alias{VEstep}
\title{VE step}
\usage{
VEstep(VE, mstep, directed, sparse, method, epsilon, fix.iter, data)
}
\arguments{
\item{VE}{Results of the previous VE step for iterative computation}

\item{mstep}{Results of the previous mstep for iterative computation}

\item{directed}{Boolean for directed (TRUE) or undirected (FALSE) case}

\item{sparse}{Boolean for sparse (TRUE) or not sparse (FALSE) case}

\item{method}{List of string. Can be "hist" for histogram method or "kernel" for kernel method}

\item{epsilon}{Threshold for the stopping criterion of VEM and fixed point iterations}

\item{fix.iter}{Maximum number of iterations of the fixed point}

\item{data}{Data same of \link[ppsbm]{mainVEM}}
}
\description{
VE step
}
\keyword{internal}
