% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_Q3}
\alias{generated_Q3}
\title{Generated graph with 50 individuals and 3 clusters}
\format{A data frame
\describe{
  \item{data}{List of 3}
  \item{z}{Latent variables}
  \item{intens}{Intensities}
}}
\usage{
generated_Q3
}
\description{
Generated graph with 50 individuals and 3 clusters
}
\keyword{datasets}
