% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_sol_hist}
\alias{generated_sol_hist}
\title{Generated solution with histogram method}
\format{List of 5 iterations of the algorithm, each one containing
\describe{
  \item{List of 8}{tau, rho, beta, logintensities.ql, best.d, J, run, converged}
}}
\usage{
generated_sol_hist
}
\description{
Generated solution with histogram method
}
\keyword{datasets}
