% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated_sol_kernel}
\alias{generated_sol_kernel}
\title{Generated solution with kernel method}
\format{Solution containing
\describe{
  \item{List of 8}{tau, logintensities.ql.ij, J, run, converged}
}}
\usage{
generated_sol_kernel
}
\description{
Generated solution with kernel method
}
\keyword{datasets}
