% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{permuteZEst}
\alias{permuteZEst}
\title{Optimal matching between 2 clusterings}
\usage{
permuteZEst(z, hat.z)
}
\arguments{
\item{z}{Matrix of size  \eqn{Q \times n}}

\item{hat.z}{Matrix of size  \eqn{Q \times n}}
}
\description{
Compute the permutation of the rows of hat.z that has to be applied to obtain the "same order" as z.
Compute optimal matching between 2 clusterings using Hungarian algorithm.
}
\references{
HUBERT, L. & ARABIE, P. (1985). Comparing partitions. J. Classif. 2, 193–218.

MATIAS, C., REBAFKA, T. & VILLERS, F. (2018).  A semiparametric extension of the stochastic block model for longitudinal networks. Biometrika. 105(3): 665-680.
}
\keyword{internal}
